+++ CHANGES.rst
@@ -4,6 +4,11 @@ Changelog
 1.3 (unreleased)
 ----------------
 
+- When showing the changelog, accept 1.7.2.1 as version.  So loose
+  version numbers instead of strict version numbers with only one or
+  two dots.
+  [maurits]
+
 - Run git pull on buildout.coredev to make sure it is up-to-date.
   [timo]
 
@@ -13,6 +18,7 @@ Changelog
 
 - Create a jenkins report based on checkPackageForUpdates.
   [gforcada]
+
 
 1.2 (2015-03-21)
 ----------------